+++ demo/components/UiHeader.vue
@@ -17,7 +17,7 @@
         :native="true"
         :class="$style.link"
       >
-        {{ item.title }}
+        {{ item.name }}
       </UiLink>
     </div>
   </header>
@@ -52,7 +52,7 @@
     border-bottom: 1px solid $colorHeather;
     padding-top: spacer(5);
     padding-bottom: spacer(3);
-    margin-bottom: spacer(5);
+    margin-bottom: spacer(6);
   }
 
   .menu {