+++ src/components/UiMarkdownViewer/index.js
@@ -1,0 +1,4 @@
+import factoryComponentPlugin from '@/utils/factoryComponentPlugin';
+import UiMarkdownViewer from './UiMarkdownViewer';
+
+export default factoryComponentPlugin(UiMarkdownViewer);